Fix a bug in index propagation of reshape.
Fix erroneous arguments to one of the calls to simplify affine
expression.
Enable reshape tests in e2e for vulkan-spirv target.

Fix the test in reshape_dropdims.mlir to actually test the code path
where unit-dimensions are dropped. Uncovers a bug in the VM, so the
test is effectively disabled.

PiperOrigin-RevId: 277743285