Tag all targets currently excluded from CI
Update the Kokoro CI to use only this tag, which allows finer-grained control. For now, I just tagged everything so I can easily check that removing those paths from the CI scripts is a no-op, but I'll go through and see if this can be less broad.

Changing the GitHub action will have to be a follow-up PR (because of our syncing infrastructure), but I still confirmed that the change will be a no-op.

Tested:
Confirmed that the bazel query commands used in the Kokoro and GitHub action CI return the same targets before and after:
$ diff -s \
  <(bazel query '//... except attr("tags", "notap", //...) except attr("tags", "nokokoro", //...) except //bindings/... except //integrations/... except //iree/hal/vulkan:dynamic_symbols_test except //iree/samples/rt:bytecode_module_api_test except //iree/samples/simple_embedding:simple_embedding_test' \
  | sort -u) \
  <(bazel query '//... except attr("tags", "nokokoro", //...)' \
  | sort -u)
...
Files /dev/fd/63 and /dev/fd/62 are identical

diff -s \
  <(bazel query '//... except attr("tags", "notap", //...) except attr("tags", "noga", //...) except //integrations/tensorflow/e2e:vulkan_conv_test except //iree/hal/vulkan:dynamic_symbols_test except //iree/samples/rt:bytecode_module_api_test except //iree/samples/simple_embedding:simple_embedding_test' \
  | sort -u) \
  <(bazel query '//... except attr("tags", "noga", //...)' \
  | sort -u)
Files /dev/fd/63 and /dev/fd/62 are identical

PiperOrigin-RevId: 286956973