Mostly rewriting the VM register allocator to allocate registers correctly.
One major issue with the previous implementation was with block arguments living outside of the block they were defined in due to the block-dominator-allows-capture behavior in MLIR. This now walks blocks in a sorted order such that all of a blocks dominators are allocated first and block live-in values (from its dominators) are reserved prior to allocating registers for the block.

Another major issue was with the branch register mapping used for br/cond_br where it was possible to end up with a cycle in the remapping DAG (making it not a DAG, obv). The simple example is a remapping of register 0->1 and then 1->0, though it's possible for much more complex graphs (such as when registers are duplicated in block argument lists). Now a feedback arc set is calculated for the remappings required by a branch. This allows us to emit a sorted DAG of most mappings (avoiding RAW hazards) and then do a swap with temp values for the remaining feedback edges that were detected.

Lots of optimization opportunity, as I disabled the reference count retain elision logic for now and as such we are performing more retains than needed at runtime and not releasing things as aggressively.

PiperOrigin-RevId: 287559518