Cleanup IREE runner tests using unfoldable constants.
This allows us to put the test inputs with the test. For now test functions in the same file are mostly using the same arguments, but this is no longer required.

Switched a few tests to use ints instead of floats where the type wasn't really relevant, since they're slightly less verbose.

PiperOrigin-RevId: 288137270