NFC: Use StructAttr for IndexComputationAttr.
Using a StructAttr cleans up a lot of the code and also allows for
generalizing the index computation attribute to represent the indices
obtained from xla_hlo operations which
1) the lowering to SPIR-V requires multiple values of an operand.
2) have multiple result values
The tests are updated to be more readable (which is possible due to
the use of the StructAttr). Note that index computation pass and
XLAToSPIRV pass need to be updated to handle cases of (1) and (2)
above. This is still TBD. Operations like xla_hlo.gather and
xla_hlo.reduce_window can be handled with this modification.

Also dropping the index_computation_attribute namespace.

PiperOrigin-RevId: 292084286