Fix OSS file reading and add tests
std::fread returns the number of objects written, not the number of bytes, which was giving inscrutable UNAVAILABLE errors in OSS.

To aid in testing, I also added a SetFileContents method, but I don't have a windows box to test this on so the windows implementation is still a TODO. I think since we don't have any windows CI some tests are certainly failing, so that's OK. I didn't really want to check this in with no tests.

Tested:
In addition to CI, new tests pass on ctest with CMake.

Closes https://github.com/google/iree/issues/687

PiperOrigin-RevId: 293622294