Start a Vulkan dialect and define attributes for target environment
To support multiple Vulkan target environments, we need a way to pass
in the configuration of the target environment. Attribute is a pragmatic
mechanism for achieving that purpose, so introducing TargetEnvAttr
for this purpose and created a Vulkan dialect to contain it as we might
gradually collect more Vulkan definitions so it's better to have the
structure set up at the beginning.

The benefits of having an attribute for representing the target environment
are that we can leverage the MLIR parser to parse it and also attach it to
various IR entities to drive conversions at various place. The former avoids
us  from shipping a JSON (or other format) parser in the compiler as long as
one queries the Vulkan environment and dump the information in the format
expected by the TargetEnvAttr. For the latter, we may consider introduce
some corresponding HAL ops so the attribute can drive the scheduling CodeGen.
At the same time, we can transform this attribute into the SPIR-V TargetEnvAttr
and use it to control SPIR-V CodeGen.

PiperOrigin-RevId: 298429559