Plumbing target environment from HAL Vulkan backend to SPIR-V CodeGen
This commit adds a command-line option to provide the Vulkan target
environment as a #vk.target_env attribute assembly and conversion
logic to get the corresponding SPIR-V target environment for driving
the CodeGen at SPIR-V side. Right now this only attaches the SPIR-V
target environment to the module sent to SPIR-V backend for compilation;
we need to thread everything up in SPIR-V CodeGen side to make sure
it is properly respected.

The default for the command-line option is SwiftShader's target
environment, but it is not fully fleshed out at the moment w.r.t.
all the supported extensions/features. That will require growing
the list of extensions and features recognized by the Vulkan backend,
which will happen in future steps. So right now it's mostly a
dummy target environment only containing the proper Vulkan version.

PiperOrigin-RevId: 298640824