Fix VMLA xla_hlo.broadcast lowering.
This was previously miscompiling, since we were blindly lowering it to IREE::VMLA::BroadcastOp, which is not correct because the latter only handles the case of a scalar broadcasting.

Since we already have the more sophisticated BroadcastInDim handled correctly, just convert xla_hlo.broadcast -> xla_hlo.broadcast_in_dim.

I'm not even bothering to implement this for interpreter-bytecode since that is going away imminently.

PiperOrigin-RevId: 299878576