Consolidate XLA to Linalg on Buffer conversion passes.
The ReductionLowering pass, the LinalgOnTensorsToBuffers pass and
LinalgToXLAOpsOnBuffers pass are all doing similar function, but being
separate passes, changes in one adversly affect other
passes. Combining them allows more streamlining of passes.
Further refactoring to move all passes into CodegenPasses folder, and
is usable by all backends from here. A function to add all passes to a
PassManager is exposed. After running these passes, it is expected
that the dispatch function contains a single op on buffers. In future
this can be relaxed to also allow for fusion with tiling of linalg on
buffer ops.
Also some cleanup to existing passes to exit more gracefully. In
particular the GPU to SPIR-V lowering requires some attributes to
exist on the gpu.func as well as the dispatch function for the
lowering. The logic to add these can be made simpler by making these
changes at gpu.module/gpu.func creation time.

PiperOrigin-RevId: 300109386