Demonstrate of check framework without lit.
This runs the compilation to a module on the host and just invokes the binary on the generated module as part of the test. Allows tests to run on devices where the compiler is not supported. Running the test does not involve relinking a C++ binary.

No CMake support yet. Future work will add wrappers to ease defining such a test in Bazel and CMake.

PiperOrigin-RevId: 301427252