Add GitHub action to automatically update submodules (#1171)
The action creates a PR that bumps the TF submodule to HEAD and the LLVM submodule to match the commit used by TF.

This automates some of the toil of keeping the submodules up to date. The build cop is still responsible for reviewing and merging these PRs (including checking that the presubmit builds pass).

Note that if we have manually bumped the LLVM submodule hash and TF hasn't this could mean the PR moves the LLVM submodule back in history.

Tested:
Tried this out in my fork. Here is an example PR it creates: https://github.com/GMNGeoffrey/iree/pull/25