Remove "DimType" from RankedShapeType
Now, only ops that actually have extents of dimensions as arguments or results care about the type used to represent dimension extents.

Having the "DimType" embedded in the type itself causes all sorts of headaches when doing conversions. For example, if one has to lower a shape.shape_of op to the ranked_shape equivalent, it forces us to prematurely take a stance on the DimType. One might suggest that in that case we should use `index` as a safe default, but that tends to cause all transitively lowered things to inherit that DimType, which eventually causes problems if there is some other op that takes a concrete position on the DimType, such as XLA which tends to use i64 as the DimType (arguably that is a bug, but we have to live with it) or VMLA which tends to use i32 (this one is more ok; at that point index has been decided to be i32).

In this CL, we do retain some amount of defaulting to IndexType, because that made the refactoring easier, and is a default that we want to push more users towards. I've left a comment in ShapeBase.td explaining some tradeoffs here so we can re-evaluate in the future.

PiperOrigin-RevId: 302589361