Migrate compare test to check framework
Vulkan Swiftshader runs of this test with non-opt builds are very slow (I'll file an issue with some traces once I can figure out how to run swiftshader in OSS). There seems to be a ton of overhead for PipelineExecutable::Create that scales with the size of the module. This doesn't come up in iree-run-mlir because it treats each test as a separate module (and actually recreates the driver per-test). After some discussion we decided to proceed with the migration and circle back to debug the swiftshader performance issues.

PiperOrigin-RevId: 306346899