Change IREE_DEFAULT_BACKENDS semantics to IREE_AVAILABLE_BACKENDS.
IREE_AVAILABLE_BACKENDS is like IREE_DEFAULT_BACKENDS but has a slightly different behavior. Rather than being a fallback, it behaves as a "mask" on the backends that are otherwise selected. This is more suitable for CI and developer workflows where certain backends aren't available.

Also, remove one test that was testing that a "disagreements" error was raised. If running with a reduced set of available backends, then it is possible to execute with just one backend, causing the test to fail since it isn't possible to compute the disagreements and raise the error.

PiperOrigin-RevId: 306347874