Fix the pattern to create a reshape op instead of removing it.
Sometimes the reshape op is gone in a canonicalize pass. In this context, we
would lose the information to adjust the filter shape. This does happen in
MobileNet model. Instead of matching if previous op being a reshape op, we do
create another reshape op right before the depthwise conv op.

The test doesn't need to be fixed because there is a folder pattern of reshape
op which makes it a no-op, i.e., the reshape op will just be removed in the
test by canonicalize pass.

Fixes https://github.com/google/iree/issues/1651

PiperOrigin-RevId: 308664401