Rework stream formation
The current stream formation implicitly assumes that the result of an arbitrary topological sort is the last operation in the block and that there exists a pre-existing place in the block where all stream ops not separated by side-effecting ops can be inserted. This breaks down in a few places.

Instead, this change dramatically simplifies the stream formation logic. It will now only form streams of adjacent streamable ops. The input pipeline is responsible for making the IR amenable for such stream formation. It turns out that the reordering the stream formation pass was doing was load bearing for dynamic shapes, however, so this change also introduces a basic hoisting pass that attempts to move all unstreamable ops as far up in their block as possible to leave clusters of streamable ops.

Allows re-enabling the unidirectional XLA HLO Unidirectional LSTM test.

Fixes https://github.com/google/iree/issues/1754

PiperOrigin-RevId: 309564075