Replacing the HAL fence with a semaphore modeled on Vulkan timeline semaphores
This gives us both compiler and runtime versions of the `hal.semaphore` and provides a basic implementation of the host semaphore (for VMLA/LLVMJIT/etc) and the Vulkan native timeline semaphore (requires the layer, extension, or Vulkan 1.2).

This change intentionally does not remove `hal.ex.submit_and_wait` as that is currently handling resource management. Future changes will add a resource timeline and use the semaphores to advance it as a way to collect resources when no longer required.

Closes https://github.com/google/iree/pull/1743

PiperOrigin-RevId: 311139790