[codegen][spirv] Add a pass to split dispatch functions
This commit adds a pass to split a dispatch function that
contains multiple Linalg ops into multiple dispatch
functions, each containing one Linalg op. This pass prepares
for lowering to GPU, where we need to tile the workload t
workgroups and workitems. If the workload involves computation
A and B, where B is dependent on A and A needs all workgroups to
complete, then we need to split A and B into different kernels
because there is no mechanism to perform cross-workgroup
synchronization within a single kernel.

Right now this pass is not integrated into the pipeline yet,
it will be used by later commits.

PiperOrigin-RevId: 311558750