[vulkan] Support multiple dispatch from one dispatch region
This commit extends the Vulkan backend to support launching
multiple kernels originated from the same dispatch region
in sequence.

The SPIR-V CodeGen pipeline incorporates the pass to split
kernels if there are multiple Linalg ops. If multiple kernels
are created, a special attribute `spv.entry_point_schedule`,
is attached to the module to indicate the scheduling of entry
points. Then the Vulkan backend queries this attribute and
dispatches entry points as scheduled.

This commit also updates various HLO to Linalg on buffers
conversion to emit a linalg.fill op to zero the output buffers
ahead of performing the real computation.

PiperOrigin-RevId: 312264954