Don't form useless streams
Right now we form streams for any streamable ops, which includes streams that are entirely shape metadata information and do no real work. This is bad on its own, but even worse, later buffer manipulation fails to populate the output of these useless streams (because it's expecting some real op to write it), which creates runtime errors.

Includes a test case that is the distillation of the repro provided by @rsuderman in https://github.com/google/iree/issues/1882#issuecomment-630445868

Helps address #1882

This is just a small fix, whereas we'd really like much more intelligent stream formation, but should help avoid painful-to-debug runtime errors.

PiperOrigin-RevId: 312306440