Add mlir-emitc submodule and extend CMake configuration
This add an EmitC dialect (forked from https://reviews.llvm.org/D76571) which is available under *Apache License v2.0 with LLVM Exceptions*. The idea is to make progress on #1173 and to utilize and extend the EmitC dialect proposed by @jpienaar for this.

Building the dialect can be enabled by `IREE_ENABLE_EMITC` in CMake. A Bazel configuration is not yet available. No additional LLVM/MLIR deps are pulled in, since `IREE_ENABLE_EMITC` results in using the `tools/init_mlir_{dialects,passes}.h` headers. No dep on the IREE target itself is introduced, instead the MLIR deps are "replicated". Also not super clean, this avoids a circular dependency and allows to avoid unnecessary dependencies (eg. on AVX512 and NVVMIR).

The final goal is to push EmitC upstream or to integrate it into IREE (at least). I am further happy to grant write access to the [mlir-emitc](https://github.com/iml130/mlir-emitc/) repo for outside collaborators as long as this doesn't prevent pushing towards MLIR or to integrate it into IREE.

Closes https://github.com/google/iree/pull/2060

PiperOrigin-RevId: 314400071