Updates `iree_vision_test_suite` to allow for divergences between dataset-backend coverage. (#2364)
Allows for the concise specification of failing `(model, dataset, backend)` tuples in the `iree_vision_test_suite` macro. This allows us to enable the `(MobileNetV2, cifar10, iree_llvmjit)` test, despite `(MobileNetV2, imagenet, iree_llvmjit)` failing.

Edit: The `(MobileNetV2, cifar10, iree_llvmjit)` config is failing again. It has been added back to the list of failing configs, but it will be easy to remove it when it starts passing again.