Fix layering and turn on layering_check (#2540)
This is a new feature in OSS Bazel (which has been available internally
for a while) that checks layering. That is, cc_* BUILD rules should
directly depend on a rule that exports each header that is `#include`'ed
in the rule's sources.

This is now available since we upgraded to Bazel 3.3.1 (#2495).

Unfortunately we can't just enable this in the .bazelrc with
`--features=layering_check` because that would enable it for all our
dependencies as well, which we are not interested in enforcing. I'm not
aware of an equivalent to `--per_file_copt` for features.

Includes adding package and license declarations to some packages that
were missing them.