Adding a size benchmark exe for measuring the total size of iree/vm. (#2561)
Adding size benchmark executable for measuring the total size of iree/vm.
When properly trimmed this should only pull in the parts of the CRT that
we use and nothing else. Today it's about ~30KB for the VM, ~30KB for
iree/base logging/status, and 400KB for random abseil/STL junk.

See issue #898 for details.