Fix help output for iree-benchmark-module (#3261)
Benchmark and Absl fight. Right now, benchmark eats the `--help` flag
and so we lose all the help information for iree-benchmark-module.

This is not elegant, but it works and includes tests. A number of better
options are not available because both Absl and Benchmark bake behavior
into the only functions callable through their public API.

Fixes https://github.com/google/iree/issues/3247