Open source `mobile_bert_squad_test` with TFLite support (#3281)
Open sources a test for MobileBert on the SQuAD task. 

Support for compiling `tf.Module`s with `tf.compat.v1.lite` was 
added to get TFLite to support compiling the model. The comparison 
`atol` is set to `1e0` because TFLite has a numerical difference of 
`~0.13` with TensorFlow on this model/input pair.

These tests don't run on the CI, so I tested them locally.