[metal] Initial bring up: properly set workgroup size (8/n) (#3422)
This commit sets up proper propagation of Metal dispatch threadgroup
sizes. Workgroup size is directly baked inside SPIR-V code, but in
Metal it's specified when dispatching workload. So when cross
compiling SPIR-V to MSL, we need to have a way to persist the
information. This is done by adding a new field in Metal's FlatBuffer
schema to plumbing it back to the Metal runtime.