Explicitly fetch tags for llvm-bazel (#3677)
We need the tags to find the matching LLVM commit and in some git
configurations (e.g. in the GitHub actions) they won't be fetched by
default. This fixes the
[Synchronize LLVM Dependents workflow](https://github.com/google/iree/actions?query=workflow%3A%22Synchronize+LLVM+Dependents%22)
which has been broken for a bit in the case where it actually needs to
fetch a new version of llvm-bazel.

Tested:
  Pushed to my fork
  https://github.com/GMNGeoffrey/iree/runs/1334664578 created
  https://github.com/GMNGeoffrey/iree/pull/60