Clarify behavior of iree-translate (#3738)
There are a couple of investigation TODOs in here that stem from
iree-translate differing from mlir-translate. They are not feasible
without reworking the tool. Instead I added additional documenation on
the differences for future readers and deleted the TODOs.

We could still reconsider the expanded role of iree-translate and reduce
it to a tool more akin to mlir-translate, creating a separate
iree-compiler tool to run full compilation passes without taking
translations to perform as command line flags. But that's larger in
scope than these TODOs.

Closes https://github.com/google/iree/issues//2958