Add file to track docker :prod digests (#3955)
This creates a source of truth for which versions of our docker images 
are reviewed and used by the CI. It is automatically updated whenever 
a contributor adds or updates an image in a PR by following the steps 
in the "Local Changes" section of the README. 

After the PR is submitted, `python3 build_tools/docker/manage_prod.py` 
must be run to update the tags on GCR. This script ensures that only the 
versions at HEAD on upstream main are pushed to `:prod`, and also 
allows a contributor to pull the images which _should_ have `:prod` tags 
without pushing them.

This simplifies updating the `:prod` tag to the point where a bot could do 
it. This partially addresses #3117.