Making iree_vm_list_resize work and switching ref cconv behavior (#4858)
The new C HAL module was the first user of this function - now there
are tests. The new shims also assumed that arguments were not retained
(as doing so requires cleanup) but the dispatch code was double-retaining
them. I'm surprised things worked as well as they did like that o_o

Fixes #4852.