Fix some issues in installed_test (#5120)
`TEST_TMPDIR` was being set based on an undefined variable `_NAME`,
which I guess we maybe happened to set in every invocation of this
function or something.

`WORKING_DIRECTORY` was being set twice in the invocation of `add_test`
and I think ignoring the argument being passed. It also was being set
to `CMAKE_BINARY_DIR` by default, which is different than the default
specified for `add_test`, which seemed weird, since that behavior
wasn't mentioned anywhere.

I hate CMake.