[CodeGen] Fuse linalg.fill and reduction linalg.generic (#5242)
This commit adjusts dispatch region formation to additionally
recognize linalg.generic as a root op and reverses the order
with which we decide fusion groups. This enables us to fuse
linalg.generic output tensors into the same group, thus we
can pull in the linalg.fill for reduction linalg.generic. This
works for both tiled cases and non-tiled cases. We actually
already treat linalg.generic as a root op (as a second step
when deciding fusion groups). This simplifies the logic by
unifying them into one.

This avoids sad dispatch regions like the following:

```mlir
  flow.executable @call_dispatch_143 attributes {sym_visibility = "private"} {
    flow.dispatch.entry @call_dispatch_143 attributes {signature = () -> tensor<f32>, workgroup_rank = 3 : index}
    module  {
      func @call_dispatch_143(%arg0: !flow.dispatch.tensor<writeonly:f32>) {
        %cst = constant 0xFF800000 : f32
        %0 = linalg.init_tensor [] : tensor<f32>
        %1 = linalg.fill(%0, %cst) : tensor<f32>, f32 -> tensor<f32>
        flow.dispatch.tensor.store %1, %arg0 : tensor<f32> -> !flow.dispatch.tensor<writeonly:f32>
        return
      }
    }
  }
```