Rework LinalgBufferize pass. (#5539)
The bufferization pass on the linalg on tensors path was an arbitrary
set of methods put in place to connect linalg on tensors to the post
tile+distribute codegeneration on CPU and GPU. This PR cleans up the
logic to better handle the tensor -> memref conversion.
In particular

Adds an analysis step to figure out which tensors can be mapped to
the same memref object. This also provides information of what
tensors are eventually mapped to the output of the dispatch
region. In such cases, the output is used to compute the values in
place reducing the need for copies and intermediate allocations.
Use this analysis step to convert operations from tensors to memrefs.
Use the readonly, readwrite and writeonly attributes to avoid
copies when possible.
All of this uses logic similar to the Tied operands/results approach
that the bufferization in the sandbox uses. Indeed this pass might
eventually subsumed by a bufferization pass in the sandbox and to be
upstreamed. Till that happens, this rewrite gives something that
is similar in spirit.

Fixes #4734
Fixes #5013