Step 4/n (n=4). Rename Conversion folder to Codegen and clean up pass names. (#6267)
The Conversion folder is renamed to Codegen (Conversion was a legacy
artefact, and Codegen is a more accurate description of the
folder). Also

CPU, SPIRV and LLVMGPU folders/passes are renamed to drop the
"LinalgTo" prefix.
Flags are renamed to use
"iree-codegen-*" for common passes
"iree-llvmcpu-*" for CPU backend passes
"iree-spirv-*" for SPIRV backend passes
"iree-llvmgpu-*" for LLVMGPU backend passes.
Filed renamed to be consistent
File name reflects name of the pass implemented in the file.
Nit: declerations in .h and .td file made alphabetical.