Start "IREE Best Practices" document. (#6541)
Trying to put some of our intuition and general IREE design goals down on [virtual] paper that we can point people at, inspired by pages like https://www.tensorflow.org/lite/performance/best_practices (for layout and general topics to cover). Some things in here like "track state within your model" are responses to existing practices for working with current-generation frameworks that we want people to untangle in their models.

I don't really have an upper limit on how long this document could be in mind. I'm starting this page on `docs/developers/`, but we could move it to `docs/website/` to have it accessible on https://google.github.io/iree/ if it is useful to enough people.