Rename Bazel repo iree_vulkan_headers to vulkan_headers (#6862)
TensorFlow used to make it impossible to name Bazel repositories the
same as theirs because they didn't use `maybe`. Now `tf_http_archive`
performs maybe-like behavior and skips the repo if one with the same
name already exists. So now we can use the more normal name for this
Bazel repo. This makes the vulkan headers Bazel repo behave like the
others in the TF build and use the IREE submodule instead of an
`http_archive` from TF. Note though, that the only usage of
`@vulkan_headers` in TF is in a part of a build that we don't use
(TFLite GPU delegate) and uses a target we don't define
("vulkan_headers_no_prototypes"):
https://github.com/tensorflow/tensorflow/blob/1b82ae90a4/tensorflow/lite/delegates/gpu/BUILD#L207