Plumb through all `TiledOpInterface` ops. (#7006)
This changes all places that use TiledOpInterface to using the
interface directly and not op-specific pattern. The external models
that implement the TiledOpInterface are also registered during
startup time.

This also has some changes to how configurations are set. For
tensor.insert_slice the lowering results in no compute ops.
So this needs to be handled. (Some of this needs some cleanup after
use of workgroup marker is dropped).

This leads to another issue on GPU. tensor.insert_slice is lowered
to a copy on bufferization. If the configuration is determined before
bufferization, this copy is missed. For now, its OK to change the
bufferization to run before configuration selection, but a better solution
is needed here.
Note: This was initially pulled together for use with
tensor.insert_slice which implements the TiledOpInterface as an
external model. This runs into some upstream bugs, so the interface is
not attached to the op as it is supposed to be till those are fixed.