Used LLVM IR for `powf` generated from MUSL for only 64-bit archs. (#7513)
Using the LLVM IR generated from powf function of MUSL for 32-bit
archs seems to result in undef. While the issue is resolved, pull in
the powf function only while compiling for 64-bit
architectures. This involves generating a new librt64.bc LLVM IR
module that is linked into the generated executables when compiled for
64-bit architectures.