Add default-flags configurations for all TFLite benchmarks (#7580)
Following up on https://github.com/google/iree/pull/7553, this adds
default flag configurations for the remaining TFLite models. Now that
we've got timestamps in the Buildkite runs, we can see what this does
to artifact transfer time via the presubmit:
https://buildkite.com/iree/iree-benchmark/builds/1447

I'm contrasting with the run at the merge base on main:
https://buildkite.com/iree/iree-benchmark/builds/1445

For this PR's run, downloads took 1:35 and 4:12 for the two RPIs,
compared to 2:06 and 3:30 before. That looks like it's basically in
the noise. We could do more detailed analysis, but that doesn't seem
worth it to me. Just spot-checking a few, it seems the RPI connected
to the Pixels consistently downloads faster than the ones connected
to the Samsungs.

As for the benchmarks themselves, the run on the RPI took
around 34-35 minutes for each phone, 6-8 min longer than
before.

Overall run latency, excluding waiting for agents, increased
from 43:12 to 49:50. I think this is acceptable given our
current limitations.