Add the fp16 <-> fp32 conversion methods to librt (along with expf for 64 bit) (#7692)
The 16 bit ops when lowered through the LLVM path use __gnu_h2f_ieee
and __gnh_f2h_ieee methods. Import an implementation of these.
Also add a expf to librt on 64-bit architectures.

The added test does not run on VMVX or SPIR-V. So enable only on CUDA
and CPU backends

Fixes #7611.