Adding alignment analysis and propagation into codegen. (#7714)
This adds stream.alignment attributes to !stream.binding and index
operands in dispatch functions and stream.values attributes to integer
operands. The alignment on bindings indicates the base binding offset
alignment while the alignment on index operands indicates the alignment
of any offset into a binding they may represent. The values are all
potential values that may be passed in the operand when knowable.

The final piece required is to use this information in the codegen
backends, however the memref.assume_alignment op doesn't seem to do
anything today. This would in theory plumb the information through
but there are likely better ways to use this information:
```
auto alignment = baseBuffer.calculateAlignment();
b.create<memref::AssumeAlignmentOp>(op->getLoc(), baseBuffer,
                                    alignment.value());
```