Remove remote execution configuration (#8021)
This is getting too difficult to maintain. To continue using it we'd
need to create an entire custom docker image (not that bad) and then
check in a ton of unreadable bazelrc and toolchain configurations
(more bad). It's already a bunch of extra overhead when bumping Bazel
versions, and would now be even more.

Instead, we can just use remote caching, which is very lightweight, and
run the builds on big machines. Unfortunately it doesn't satisfy my
vision of people without 96-core machines being able to build anything
with LLVM or TF dependencies, but that never came to fruition with RBE
remote execution either.

Fixes https://github.com/google/iree/issues/7419