Initial commit for Partitionable Loop Interface. (#8033)
This adds the initial implementation of interface to specify which
loops are partitionable. This change just adds the external models for
all operations this is needed for.
A test pass is added that just generates a util.unfoldable_const
with values that show which loops of the operation are partitionable.